conlict fixed, and refactoring dashboard module
