+++ lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_slicing_project/core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_slicing_project/module/dashboard/view/dashboard5_view.dart';
 import 'package:flutter_slicing_project/utils/color_lib.dart';
 
 void main() {
@@ -18,7 +17,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSwatch(primarySwatch: Colors.green)
             .copyWith(background: ColorLib.darkGray),
       ),
-      home: const Dashboard5View(),
+      home: const DashboardView(),
       navigatorKey: Get.navigatorKey,
     );
   }
