+++ lib/module/dashboard/controller/edashboard5/dashboard5_controller.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_slicing_project/core.dart';
 
-import '../../../state_util.dart';
-import '../view/dashboard5_view.dart';
+import '../../../../state_util.dart';
+import '../../view/edashboard5/dashboard5_view.dart';
 
 class Dashboard5Controller extends State<Dashboard5View>
     implements MvcController {
